+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 #### Changed
 - Using `giveaway` command requires "manage guild" permission.
 - Increased `meme`'s cooldown from 4s to 6s.
+- `giveaway`'s help is now more precise.
 
 ## Released